requirements.txt
aioitertools==0.12.0
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arize-otel==0.8.1
arize-phoenix==9.5.0
arize-phoenix-client==1.6.0
arize-phoenix-evals==0.20.6
arize-phoenix-otel==0.9.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.5.2
blinker==1.9.0
boto3==1.38.19
botocore==1.38.19
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.40.1
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphql-core==3.2.6
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
Mako==1.3.10
MarkupSafe==3.0.2
narwhals==1.39.1
numpy==2.2.6
openai==1.79.0
openinference-instrumentation==0.1.29
openinference-instrumentation-langchain==0.1.43
openinference-semantic-conventions==0.1.17
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.0
rsa==4.9.1
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake==1.5.0
snowflake-connector-python==3.15.0
snowflake._legacy==1.0.0
snowflake.core==1.5.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlean.py==3.49.1
starlette==0.46.2
strawberry-graphql==0.269.0
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0